refactor: move commenting to "success" lifecycle
