+++ lib/definitions/constants.js
@@ -1,0 +1,3 @@
+const RELEASE_NAME = 'GitLab release';
+
+module.exports = {RELEASE_NAME};
