+++ test/publish.test.js
@@ -45,60 +45,6 @@ test.serial('Publish a release', async t => {
   t.true(gitlab.isDone());
 });
 
-test.serial('Publish a release and post comments', async t => {
-  const owner = 'test_user';
-  const repo = 'test_repo';
-  const env = {GITLAB_TOKEN: 'gitlab_token'};
-  const pluginConfig = {postComments: true};
-  const nextRelease = {gitHead: '123', gitTag: 'v1.0.0', notes: 'Test release note body'};
-  const options = {repositoryUrl: `https://gitlab.com/${owner}/${repo}.git`};
-  const encodedRepoId = encodeURIComponent(`${owner}/${repo}`);
-  const commits = [{hash: 'abcdef'}, {hash: 'fedcba'}];
-  const gitlab = authenticate(env)
-    .post(`/projects/${encodedRepoId}/releases`, {
-      tag_name: nextRelease.gitTag,
-      description: nextRelease.notes,
-      assets: {
-        links: [],
-      },
-    })
-    .reply(200)
-    .get(`/projects/${encodedRepoId}/repository/commits/abcdef/merge_requests`)
-    .reply(200, [
-      {project_id: 100, iid: 1, state: 'merged'},
-      {project_id: 200, iid: 2, state: 'closed'},
-      {project_id: 300, iid: 3, state: 'merged'},
-    ])
-    .get(`/projects/${encodedRepoId}/repository/commits/fedcba/merge_requests`)
-    .reply(200, [{project_id: 100, iid: 1, state: 'merged'}])
-    .get(`/projects/100/merge_requests/1/closes_issues`)
-    .reply(200, [
-      {project_id: 100, iid: 11, state: 'closed'},
-      {project_id: 100, iid: 12, state: 'open'},
-      {project_id: 100, iid: 13, state: 'closed'},
-    ])
-    .get(`/projects/300/merge_requests/3/closes_issues`)
-    .reply(200, [])
-    .post(`/projects/100/merge_requests/1/notes`, {
-      body:
-        ':tada: This MR is included in version v1.0.0 :tada:\n\nThe release is available on [Gitlab Release](https://gitlab.com/test_user%2Ftest_repo/-/releases/v1.0.0)',
-    })
-    .reply(200)
-    .post(`/projects/300/merge_requests/3/notes`)
-    .reply(200)
-    .post(`/projects/100/issues/11/notes`, {
-      body:
-        ':tada: This issue has been resolved in version v1.0.0 :tada:\n\nThe release is available on [Gitlab Release](https://gitlab.com/test_user%2Ftest_repo/-/releases/v1.0.0)',
-    })
-    .reply(200)
-    .post(`/projects/100/issues/13/notes`)
-    .reply(200);
-
-  await publish(pluginConfig, {env, options, nextRelease, logger: t.context.logger, commits});
-
-  t.true(gitlab.isDone());
-});
-
 test.serial('Publish a release with assets', async t => {
   const cwd = 'test/fixtures/files';
   const owner = 'test_user';
